fix(mafia): fix bug with anarch bomb exploding in postgame (#1121)

+++ Games/types/Mafia/items/Timebomb.js
@@ -24,6 +24,10 @@ module.exports = class Timebomb extends Item {
         // bomb detonates between 10 and 30 seconds
         let toDetonate = Random.randInt(10000, 30000);
         this.timer = setTimeout(() => {
+          if (this.game.finished) {
+            return;
+          }
+
           this.drop();
           if (!this.holder.alive) {
             return;
